Tools/mmaps_generator: Move TileBuilder to its own file
