+++ src/tools/mmaps_generator/MapBuilder.cpp
@@ -16,21 +16,16 @@
  */
 
 #include "MapBuilder.h"
-#include "IntermediateValues.h"
 #include "Log.h"
 #include "MapDefines.h"
 #include "MapTree.h"
 #include "MapUtils.h"
 #include "Memory.h"
 #include "MMapDefines.h"
-#include "ModelInstance.h"
 #include "PathCommon.h"
 #include "StringConvert.h"
 #include "StringFormat.h"
-#include <DetourNavMesh.h>
-#include <DetourNavMeshBuilder.h>
 #include <boost/filesystem/directory.hpp>
-#include <climits>
 
 namespace FileExtensions
 {
@@ -40,36 +35,33 @@ static boost::filesystem::path vmtile = ".vmtile";
 
 namespace MMAP
 {
-    TileBuilder::TileBuilder(MapBuilder* mapBuilder, bool skipLiquid, bool bigBaseUnit, bool debugOutput) :
-        m_bigBaseUnit(bigBaseUnit),
-        m_debugOutput(debugOutput),
+    MapTileBuilder::MapTileBuilder(MapBuilder* mapBuilder, Optional<float> maxWalkableAngle, Optional<float> maxWalkableAngleNotSteep,
+        bool skipLiquid, bool bigBaseUnit, bool debugOutput, std::vector<OffMeshData> const* offMeshConnections) :
+        TileBuilder(maxWalkableAngle, maxWalkableAngleNotSteep, skipLiquid, bigBaseUnit, debugOutput, offMeshConnections),
         m_mapBuilder(mapBuilder),
-        m_terrainBuilder(nullptr),
-        m_workerThread(&TileBuilder::WorkerThread, this),
-        m_rcContext(nullptr)
+        m_workerThread(&MapTileBuilder::WorkerThread, this)
     {
-        m_terrainBuilder = new TerrainBuilder(skipLiquid);
-        m_rcContext = new rcContext(false);
     }
 
-    TileBuilder::~TileBuilder()
+    MapTileBuilder::~MapTileBuilder()
     {
         WaitCompletion();
-
-        delete m_terrainBuilder;
-        delete m_rcContext;
     }
 
-    void TileBuilder::WaitCompletion()
+    void MapTileBuilder::WaitCompletion()
     {
         if (m_workerThread.joinable())
             m_workerThread.join();
     }
 
+    void MapTileBuilder::OnTileDone()
+    {
+        ++m_mapBuilder->m_totalTilesProcessed;
+    }
+
     MapBuilder::MapBuilder(Optional<float> maxWalkableAngle, Optional<float> maxWalkableAngleNotSteep, bool skipLiquid,
         bool skipContinents, bool skipJunkMaps, bool skipBattlegrounds,
         bool debugOutput, bool bigBaseUnit, int mapid, char const* offMeshFilePath, unsigned int threads) :
-        m_terrainBuilder     (nullptr),
         m_debugOutput        (debugOutput),
         m_threads            (threads),
         m_skipContinents     (skipContinents),
@@ -82,12 +74,8 @@ namespace MMAP
         m_mapid              (mapid),
         m_totalTiles         (0u),
         m_totalTilesProcessed(0u),
-        m_rcContext          (nullptr),
         _cancelationToken    (false)
     {
-        m_terrainBuilder = new TerrainBuilder(skipLiquid);
-
-        m_rcContext = new rcContext(false);
 
         // At least 1 thread is needed
         m_threads = std::max(1u, m_threads);
@@ -109,9 +97,6 @@ namespace MMAP
 
         m_tileBuilders.clear();
         m_tiles.clear();
-
-        delete m_terrainBuilder;
-        delete m_rcContext;
     }
 
     /**************************************************************************/
@@ -238,7 +223,7 @@ namespace MMAP
 
     /**************************************************************************/
 
-    void TileBuilder::WorkerThread()
+    void MapTileBuilder::WorkerThread()
     {
         while (true)
         {
@@ -269,7 +254,8 @@ namespace MMAP
 
         for (unsigned int i = 0; i < m_threads; ++i)
         {
-            m_tileBuilders.push_back(new TileBuilder(this, m_skipLiquid, m_bigBaseUnit, m_debugOutput));
+            m_tileBuilders.push_back(new MapTileBuilder(this, m_maxWalkableAngle, m_maxWalkableAngleNotSteep,
+                m_skipLiquid, m_bigBaseUnit, m_debugOutput, &m_offMeshConnections));
         }
 
         if (mapID)
@@ -376,10 +362,11 @@ namespace MMAP
         TerrainBuilder::cleanVertices(data.solidVerts, data.solidTris);
         // get bounds of current tile
         float bmin[3], bmax[3];
-        getTileBounds(tileX, tileY, data.solidVerts.getCArray(), data.solidVerts.size() / 3, bmin, bmax);
+        TileBuilder::getTileBounds(tileX, tileY, data.solidVerts.getCArray(), data.solidVerts.size() / 3, bmin, bmax);
 
         // build navmesh tile
-        TileBuilder tileBuilder = TileBuilder(this, m_skipLiquid, m_bigBaseUnit, m_debugOutput);
+        MapTileBuilder tileBuilder(this, m_maxWalkableAngle, m_maxWalkableAngleNotSteep,
+            m_skipLiquid, m_bigBaseUnit, m_debugOutput, &m_offMeshConnections);
         tileBuilder.buildMoveMapTile(mapId, tileX, tileY, data, bmin, bmax, navMesh);
         fclose(file);
     }
@@ -397,7 +384,8 @@ namespace MMAP
 
         // ToDo: delete the old tile as the user clearly wants to rebuild it
 
-        TileBuilder tileBuilder = TileBuilder(this, m_skipLiquid, m_bigBaseUnit, m_debugOutput);
+        MapTileBuilder tileBuilder(this, m_maxWalkableAngle, m_maxWalkableAngleNotSteep,
+            m_skipLiquid, m_bigBaseUnit, m_debugOutput, &m_offMeshConnections);
         tileBuilder.buildTile(mapID, tileX, tileY, navMesh);
         dtFreeNavMesh(navMesh);
 
@@ -445,59 +433,6 @@ namespace MMAP
     }
 
     /**************************************************************************/
-    void TileBuilder::buildTile(uint32 mapID, uint32 tileX, uint32 tileY, dtNavMesh* navMesh)
-    {
-        if(shouldSkipTile(mapID, tileX, tileY))
-        {
-            ++m_mapBuilder->m_totalTilesProcessed;
-            return;
-        }
-
-        TC_LOG_INFO("maps.mmapgen", "{}% [Map {:04}] Building tile [{:02},{:02}]", m_mapBuilder->currentPercentageDone(), mapID, tileX, tileY);
-
-        MeshData meshData;
-
-        // get heightmap data
-        m_terrainBuilder->loadMap(mapID, tileX, tileY, meshData);
-
-        // get model data
-        m_terrainBuilder->loadVMap(mapID, tileY, tileX, meshData);
-
-        // if there is no data, give up now
-        if (!meshData.solidVerts.size() && !meshData.liquidVerts.size())
-        {
-            ++m_mapBuilder->m_totalTilesProcessed;
-            return;
-        }
-
-        // remove unused vertices
-        TerrainBuilder::cleanVertices(meshData.solidVerts, meshData.solidTris);
-        TerrainBuilder::cleanVertices(meshData.liquidVerts, meshData.liquidTris);
-
-        // gather all mesh data for final data check, and bounds calculation
-        G3D::Array<float> allVerts;
-        allVerts.append(meshData.liquidVerts);
-        allVerts.append(meshData.solidVerts);
-
-        if (!allVerts.size())
-        {
-            ++m_mapBuilder->m_totalTilesProcessed;
-            return;
-        }
-
-        // get bounds of current tile
-        float bmin[3], bmax[3];
-        m_mapBuilder->getTileBounds(tileX, tileY, allVerts.getCArray(), allVerts.size() / 3, bmin, bmax);
-
-        m_terrainBuilder->loadOffMeshConnections(mapID, tileX, tileY, meshData, m_mapBuilder->m_offMeshConnections);
-
-        // build navmesh tile
-        buildMoveMapTile(mapID, tileX, tileY, meshData, bmin, bmax, navMesh);
-
-        ++m_mapBuilder->m_totalTilesProcessed;
-    }
-
-    /**************************************************************************/
     void MapBuilder::buildNavMesh(uint32 mapID, dtNavMesh* &navMesh)
     {
         // if map has a parent we use that to generate dtNavMeshParams - worldserver will load all missing tiles from that map
@@ -542,7 +477,7 @@ namespace MMAP
 
         // use Max because '32 - tileX' is negative for values over 32
         float bmin[3], bmax[3];
-        getTileBounds(tileXMax, tileYMax, nullptr, 0, bmin, bmax);
+        TileBuilder::getTileBounds(tileXMax, tileYMax, nullptr, 0, bmin, bmax);
 
         /***       now create the navmesh       ***/
 
@@ -577,361 +512,6 @@ namespace MMAP
         // now that we know navMesh params are valid, we can write them to file
         fwrite(&navMeshParams, sizeof(dtNavMeshParams), 1, file);
         fclose(file);
-    }
-
-    /**************************************************************************/
-    void TileBuilder::buildMoveMapTile(uint32 mapID, uint32 tileX, uint32 tileY,
-        MeshData &meshData, float bmin[3], float bmax[3],
-        dtNavMesh* navMesh)
-    {
-        // console output
-        std::string tileString = Trinity::StringFormat("[Map {:04}] [{:02},{:02}]: ", mapID, tileX, tileY);
-        TC_LOG_INFO("maps.mmapgen", "{} Building movemap tile...", tileString);
-
-        IntermediateValues iv;
-
-        float* tVerts = meshData.solidVerts.getCArray();
-        int tVertCount = meshData.solidVerts.size() / 3;
-        int* tTris = meshData.solidTris.getCArray();
-        int tTriCount = meshData.solidTris.size() / 3;
-
-        float* lVerts = meshData.liquidVerts.getCArray();
-        int lVertCount = meshData.liquidVerts.size() / 3;
-        int* lTris = meshData.liquidTris.getCArray();
-        int lTriCount = meshData.liquidTris.size() / 3;
-        uint8* lTriFlags = meshData.liquidType.getCArray();
-
-        const TileConfig tileConfig = TileConfig(m_bigBaseUnit);
-        int TILES_PER_MAP = tileConfig.TILES_PER_MAP;
-        float BASE_UNIT_DIM = tileConfig.BASE_UNIT_DIM;
-        rcConfig config = m_mapBuilder->GetMapSpecificConfig(mapID, bmin, bmax, tileConfig);
-
-        // this sets the dimensions of the heightfield - should maybe happen before border padding
-        rcCalcGridSize(config.bmin, config.bmax, config.cs, &config.width, &config.height);
-
-        // allocate subregions : tiles
-        Tile* tiles = new Tile[TILES_PER_MAP * TILES_PER_MAP];
-
-        // Initialize per tile config.
-        rcConfig tileCfg = config;
-        tileCfg.width = config.tileSize + config.borderSize*2;
-        tileCfg.height = config.tileSize + config.borderSize*2;
-
-        // merge per tile poly and detail meshes
-        rcPolyMesh** pmmerge = new rcPolyMesh*[TILES_PER_MAP * TILES_PER_MAP];
-        rcPolyMeshDetail** dmmerge = new rcPolyMeshDetail*[TILES_PER_MAP * TILES_PER_MAP];
-        int nmerge = 0;
-        // build all tiles
-        for (int y = 0; y < TILES_PER_MAP; ++y)
-        {
-            for (int x = 0; x < TILES_PER_MAP; ++x)
-            {
-                Tile& tile = tiles[x + y * TILES_PER_MAP];
-
-                // Calculate the per tile bounding box.
-                tileCfg.bmin[0] = config.bmin[0] + x * float(config.tileSize * config.cs);
-                tileCfg.bmin[2] = config.bmin[2] + y * float(config.tileSize * config.cs);
-                tileCfg.bmax[0] = config.bmin[0] + (x + 1) * float(config.tileSize * config.cs);
-                tileCfg.bmax[2] = config.bmin[2] + (y + 1) * float(config.tileSize * config.cs);
-
-                tileCfg.bmin[0] -= tileCfg.borderSize * tileCfg.cs;
-                tileCfg.bmin[2] -= tileCfg.borderSize * tileCfg.cs;
-                tileCfg.bmax[0] += tileCfg.borderSize * tileCfg.cs;
-                tileCfg.bmax[2] += tileCfg.borderSize * tileCfg.cs;
-
-                // build heightfield
-                tile.solid = rcAllocHeightfield();
-                if (!tile.solid || !rcCreateHeightfield(m_rcContext, *tile.solid, tileCfg.width, tileCfg.height, tileCfg.bmin, tileCfg.bmax, tileCfg.cs, tileCfg.ch))
-                {
-                    TC_LOG_ERROR("maps.mmapgen", "{} Failed building heightfield!", tileString);
-                    continue;
-                }
-
-                // mark all walkable tiles, both liquids and solids
-
-                /* we want to have triangles with slope less than walkableSlopeAngleNotSteep (<= 55) to have NAV_AREA_GROUND
-                 * and with slope between walkableSlopeAngleNotSteep and walkableSlopeAngle (55 < .. <= 70) to have NAV_AREA_GROUND_STEEP.
-                 * we achieve this using recast API: memset everything to NAV_AREA_GROUND_STEEP, call rcClearUnwalkableTriangles with 70 so
-                 * any area above that will get RC_NULL_AREA (unwalkable), then call rcMarkWalkableTriangles with 55 to set NAV_AREA_GROUND
-                 * on anything below 55 . Players and idle Creatures can use NAV_AREA_GROUND, while Creatures in combat can use NAV_AREA_GROUND_STEEP.
-                 */
-                unsigned char* triFlags = new unsigned char[tTriCount];
-                memset(triFlags, NAV_AREA_GROUND_STEEP, tTriCount*sizeof(unsigned char));
-                rcClearUnwalkableTriangles(m_rcContext, tileCfg.walkableSlopeAngle, tVerts, tVertCount, tTris, tTriCount, triFlags);
-                rcMarkWalkableTriangles(m_rcContext, tileCfg.walkableSlopeAngleNotSteep, tVerts, tVertCount, tTris, tTriCount, triFlags, NAV_AREA_GROUND);
-                rcRasterizeTriangles(m_rcContext, tVerts, tVertCount, tTris, triFlags, tTriCount, *tile.solid, config.walkableClimb);
-                delete[] triFlags;
-
-                rcFilterLowHangingWalkableObstacles(m_rcContext, config.walkableClimb, *tile.solid);
-                rcFilterLedgeSpans(m_rcContext, tileCfg.walkableHeight, tileCfg.walkableClimb, *tile.solid);
-                rcFilterWalkableLowHeightSpans(m_rcContext, tileCfg.walkableHeight, *tile.solid);
-
-                // add liquid triangles
-                rcRasterizeTriangles(m_rcContext, lVerts, lVertCount, lTris, lTriFlags, lTriCount, *tile.solid, config.walkableClimb);
-
-                // compact heightfield spans
-                tile.chf = rcAllocCompactHeightfield();
-                if (!tile.chf || !rcBuildCompactHeightfield(m_rcContext, tileCfg.walkableHeight, tileCfg.walkableClimb, *tile.solid, *tile.chf))
-                {
-                    TC_LOG_ERROR("maps.mmapgen", "{} Failed compacting heightfield!", tileString);
-                    continue;
-                }
-
-                // build polymesh intermediates
-                if (!rcErodeWalkableArea(m_rcContext, config.walkableRadius, *tile.chf))
-                {
-                    TC_LOG_ERROR("maps.mmapgen", "{} Failed eroding area!", tileString);
-                    continue;
-                }
-
-                if (!rcMedianFilterWalkableArea(m_rcContext, *tile.chf))
-                {
-                    TC_LOG_ERROR("maps.mmapgen", "{} Failed filtering area!", tileString);
-                    continue;
-                }
-
-                if (!rcBuildDistanceField(m_rcContext, *tile.chf))
-                {
-                    TC_LOG_ERROR("maps.mmapgen", "{} Failed building distance field!", tileString);
-                    continue;
-                }
-
-                if (!rcBuildRegions(m_rcContext, *tile.chf, tileCfg.borderSize, tileCfg.minRegionArea, tileCfg.mergeRegionArea))
-                {
-                    TC_LOG_ERROR("maps.mmapgen", "{} Failed building regions!", tileString);
-                    continue;
-                }
-
-                tile.cset = rcAllocContourSet();
-                if (!tile.cset || !rcBuildContours(m_rcContext, *tile.chf, tileCfg.maxSimplificationError, tileCfg.maxEdgeLen, *tile.cset))
-                {
-                    TC_LOG_ERROR("maps.mmapgen", "{} Failed building contours!", tileString);
-                    continue;
-                }
-
-                // build polymesh
-                tile.pmesh = rcAllocPolyMesh();
-                if (!tile.pmesh || !rcBuildPolyMesh(m_rcContext, *tile.cset, tileCfg.maxVertsPerPoly, *tile.pmesh))
-                {
-                    TC_LOG_ERROR("maps.mmapgen", "{} Failed building polymesh!", tileString);
-                    continue;
-                }
-
-                tile.dmesh = rcAllocPolyMeshDetail();
-                if (!tile.dmesh || !rcBuildPolyMeshDetail(m_rcContext, *tile.pmesh, *tile.chf, tileCfg.detailSampleDist, tileCfg.detailSampleMaxError, *tile.dmesh))
-                {
-                    TC_LOG_ERROR("maps.mmapgen", "{} Failed building polymesh detail!", tileString);
-                    continue;
-                }
-
-                // free those up
-                // we may want to keep them in the future for debug
-                // but right now, we don't have the code to merge them
-                rcFreeHeightField(tile.solid);
-                tile.solid = nullptr;
-                rcFreeCompactHeightfield(tile.chf);
-                tile.chf = nullptr;
-                rcFreeContourSet(tile.cset);
-                tile.cset = nullptr;
-
-                pmmerge[nmerge] = tile.pmesh;
-                dmmerge[nmerge] = tile.dmesh;
-                nmerge++;
-            }
-        }
-
-        iv.polyMesh = rcAllocPolyMesh();
-        if (!iv.polyMesh)
-        {
-            TC_LOG_ERROR("maps.mmapgen", "{} alloc iv.polyMesh FAILED!", tileString);
-            delete[] pmmerge;
-            delete[] dmmerge;
-            delete[] tiles;
-            return;
-        }
-        rcMergePolyMeshes(m_rcContext, pmmerge, nmerge, *iv.polyMesh);
-
-        iv.polyMeshDetail = rcAllocPolyMeshDetail();
-        if (!iv.polyMeshDetail)
-        {
-            TC_LOG_ERROR("maps.mmapgen", "{} alloc m_dmesh FAILED!", tileString);
-            delete[] pmmerge;
-            delete[] dmmerge;
-            delete[] tiles;
-            return;
-        }
-        rcMergePolyMeshDetails(m_rcContext, dmmerge, nmerge, *iv.polyMeshDetail);
-
-        // free things up
-        delete[] pmmerge;
-        delete[] dmmerge;
-        delete[] tiles;
-
-        // set polygons as walkable
-        // TODO: special flags for DYNAMIC polygons, ie surfaces that can be turned on and off
-        for (int i = 0; i < iv.polyMesh->npolys; ++i)
-        {
-            if (uint8 area = iv.polyMesh->areas[i] & NAV_AREA_ALL_MASK)
-            {
-                if (area >= NAV_AREA_MIN_VALUE)
-                    iv.polyMesh->flags[i] = 1 << (NAV_AREA_MAX_VALUE - area);
-                else
-                    iv.polyMesh->flags[i] = NAV_GROUND; // TODO: these will be dynamic in future
-            }
-        }
-
-        // setup mesh parameters
-        dtNavMeshCreateParams params;
-        memset(&params, 0, sizeof(params));
-        params.verts = iv.polyMesh->verts;
-        params.vertCount = iv.polyMesh->nverts;
-        params.polys = iv.polyMesh->polys;
-        params.polyAreas = iv.polyMesh->areas;
-        params.polyFlags = iv.polyMesh->flags;
-        params.polyCount = iv.polyMesh->npolys;
-        params.nvp = iv.polyMesh->nvp;
-        params.detailMeshes = iv.polyMeshDetail->meshes;
-        params.detailVerts = iv.polyMeshDetail->verts;
-        params.detailVertsCount = iv.polyMeshDetail->nverts;
-        params.detailTris = iv.polyMeshDetail->tris;
-        params.detailTriCount = iv.polyMeshDetail->ntris;
-
-        params.offMeshConVerts = meshData.offMeshConnections.getCArray();
-        params.offMeshConCount = meshData.offMeshConnections.size()/6;
-        params.offMeshConRad = meshData.offMeshConnectionRads.getCArray();
-        params.offMeshConDir = meshData.offMeshConnectionDirs.getCArray();
-        params.offMeshConAreas = meshData.offMeshConnectionsAreas.getCArray();
-        params.offMeshConFlags = meshData.offMeshConnectionsFlags.getCArray();
-
-        params.walkableHeight = BASE_UNIT_DIM*config.walkableHeight;    // agent height
-        params.walkableRadius = BASE_UNIT_DIM*config.walkableRadius;    // agent radius
-        params.walkableClimb = BASE_UNIT_DIM*config.walkableClimb;      // keep less that walkableHeight (aka agent height)!
-        params.tileX = (((bmin[0] + bmax[0]) / 2) - navMesh->getParams()->orig[0]) / GRID_SIZE;
-        params.tileY = (((bmin[2] + bmax[2]) / 2) - navMesh->getParams()->orig[2]) / GRID_SIZE;
-        rcVcopy(params.bmin, bmin);
-        rcVcopy(params.bmax, bmax);
-        params.cs = config.cs;
-        params.ch = config.ch;
-        params.tileLayer = 0;
-        params.buildBvTree = true;
-
-        // will hold final navmesh
-        unsigned char* navData = nullptr;
-        int navDataSize = 0;
-
-        do
-        {
-            // these values are checked within dtCreateNavMeshData - handle them here
-            // so we have a clear error message
-            if (params.nvp > DT_VERTS_PER_POLYGON)
-            {
-                TC_LOG_ERROR("maps.mmapgen", "{} Invalid verts-per-polygon value!", tileString);
-                break;
-            }
-            if (params.vertCount >= 0xffff)
-            {
-                TC_LOG_ERROR("maps.mmapgen", "{} Too many vertices!", tileString);
-                break;
-            }
-            if (!params.vertCount || !params.verts)
-            {
-                // occurs mostly when adjacent tiles have models
-                // loaded but those models don't span into this tile
-
-                // message is an annoyance
-                //TC_LOG_ERROR("maps.mmapgen", "{} No vertices to build tile!", tileString);
-                break;
-            }
-            if (!params.polyCount || !params.polys)
-            {
-                // we have flat tiles with no actual geometry - don't build those, its useless
-                // keep in mind that we do output those into debug info
-                TC_LOG_ERROR("maps.mmapgen", "{} No polygons to build on tile!", tileString);
-                break;
-            }
-            if (!params.detailMeshes || !params.detailVerts || !params.detailTris)
-            {
-                TC_LOG_ERROR("maps.mmapgen", "{} No detail mesh to build tile!", tileString);
-                break;
-            }
-
-            TC_LOG_DEBUG("maps.mmapgen", "{} Building navmesh tile...", tileString);
-            if (!dtCreateNavMeshData(&params, &navData, &navDataSize))
-            {
-                TC_LOG_ERROR("maps.mmapgen", "{} Failed building navmesh tile!", tileString);
-                break;
-            }
-
-            dtTileRef tileRef = 0;
-            TC_LOG_DEBUG("maps.mmapgen", "{} Adding tile to navmesh...", tileString);
-            // DT_TILE_FREE_DATA tells detour to unallocate memory when the tile
-            // is removed via removeTile()
-            dtStatus dtResult = navMesh->addTile(navData, navDataSize, DT_TILE_FREE_DATA, 0, &tileRef);
-            if (!tileRef || !dtStatusSucceed(dtResult))
-            {
-                TC_LOG_ERROR("maps.mmapgen", "{} Failed adding tile to navmesh!", tileString);
-                break;
-            }
-
-            // file output
-            std::string fileName = Trinity::StringFormat("mmaps/{:04}{:02}{:02}.mmtile", mapID, tileY, tileX);
-            FILE* file = fopen(fileName.c_str(), "wb");
-            if (!file)
-            {
-                TC_LOG_ERROR("maps.mmapgen", "{}: [Map {:04}] Failed to open {} for writing!", strerror(errno), mapID, fileName);
-                navMesh->removeTile(tileRef, nullptr, nullptr);
-                break;
-            }
-
-            TC_LOG_DEBUG("maps.mmapgen", "{} Writing to file...", tileString);
-
-            // write header
-            MmapTileHeader header;
-            header.usesLiquids = m_terrainBuilder->usesLiquids();
-            header.size = uint32(navDataSize);
-            fwrite(&header, sizeof(MmapTileHeader), 1, file);
-
-            // write data
-            fwrite(navData, sizeof(unsigned char), navDataSize, file);
-            fclose(file);
-
-            // now that tile is written to disk, we can unload it
-            navMesh->removeTile(tileRef, nullptr, nullptr);
-        }
-        while (false);
-
-        if (m_debugOutput)
-        {
-            // restore padding so that the debug visualization is correct
-            for (int i = 0; i < iv.polyMesh->nverts; ++i)
-            {
-                unsigned short* v = &iv.polyMesh->verts[i*3];
-                v[0] += (unsigned short)config.borderSize;
-                v[2] += (unsigned short)config.borderSize;
-            }
-
-            iv.generateObjFile(mapID, tileX, tileY, meshData);
-            iv.writeIV(mapID, tileX, tileY);
-        }
-    }
-
-    /**************************************************************************/
-    void MapBuilder::getTileBounds(uint32 tileX, uint32 tileY, float* verts, int vertCount, float* bmin, float* bmax) const
-    {
-        // this is for elevation
-        if (verts && vertCount)
-            rcCalcBounds(verts, vertCount, bmin, bmax);
-        else
-        {
-            bmin[1] = FLT_MIN;
-            bmax[1] = FLT_MAX;
-        }
-
-        // this is for width and depth
-        bmax[0] = (32 - int(tileX)) * GRID_SIZE;
-        bmax[2] = (32 - int(tileY)) * GRID_SIZE;
-        bmin[0] = bmax[0] - GRID_SIZE;
-        bmin[2] = bmax[2] - GRID_SIZE;
     }
 
     /**************************************************************************/
@@ -1020,7 +600,7 @@ namespace MMAP
     }
 
     /**************************************************************************/
-    bool TileBuilder::shouldSkipTile(uint32 mapID, uint32 tileX, uint32 tileY) const
+    bool MapTileBuilder::shouldSkipTile(uint32 mapID, uint32 tileX, uint32 tileY) const
     {
         std::string fileName = Trinity::StringFormat("mmaps/{:04}{:02}{:02}.mmtile", mapID, tileY, tileX);
         FILE* file = fopen(fileName.c_str(), "rb");
@@ -1039,58 +619,12 @@ namespace MMAP
         if (header.mmapVersion != MMAP_VERSION)
             return false;
 
-        if (m_debugOutput)
-            return false;
-
-        return true;
+        return TileBuilder::shouldSkipTile(mapID, tileX, tileY);
     }
 
-    rcConfig MapBuilder::GetMapSpecificConfig(uint32 mapID, float bmin[3], float bmax[3], const TileConfig &tileConfig) const
+    std::string MapTileBuilder::GetProgressText() const
     {
-        rcConfig config;
-        memset(&config, 0, sizeof(rcConfig));
-
-        rcVcopy(config.bmin, bmin);
-        rcVcopy(config.bmax, bmax);
-
-        config.maxVertsPerPoly = DT_VERTS_PER_POLYGON;
-        config.cs = tileConfig.BASE_UNIT_DIM;
-        config.ch = tileConfig.BASE_UNIT_DIM;
-        // Keeping these 2 slope angles the same reduces a lot the number of polys.
-        // 55 should be the minimum, maybe 70 is ok (keep in mind blink uses mmaps), 85 is too much for players
-        config.walkableSlopeAngle = m_maxWalkableAngle ? *m_maxWalkableAngle : 55;
-        config.walkableSlopeAngleNotSteep = m_maxWalkableAngleNotSteep ? *m_maxWalkableAngleNotSteep : 55;
-        config.tileSize = tileConfig.VERTEX_PER_TILE;
-        config.walkableRadius = m_bigBaseUnit ? 1 : 2;
-        config.borderSize = config.walkableRadius + 3;
-        config.maxEdgeLen = tileConfig.VERTEX_PER_TILE + 1;        // anything bigger than tileSize
-        config.walkableHeight = m_bigBaseUnit ? 3 : 6;
-        // a value >= 3|6 allows npcs to walk over some fences
-        // a value >= 4|8 allows npcs to walk over all fences
-        config.walkableClimb = m_bigBaseUnit ? 3 : 6;
-        config.minRegionArea = rcSqr(60);
-        config.mergeRegionArea = rcSqr(50);
-        config.maxSimplificationError = 1.8f;           // eliminates most jagged edges (tiny polygons)
-        config.detailSampleDist = config.cs * 16;
-        config.detailSampleMaxError = config.ch * 1;
-
-        switch (mapID)
-        {
-            // Blade's Edge Arena
-            case 562:
-                // This allows to walk on the ropes to the pillars
-                config.walkableRadius = 0;
-                break;
-            // Blackfathom Deeps
-            case 48:
-                // Reduce the chance to have underground levels
-                config.ch *= 2;
-                break;
-            default:
-                break;
-        }
-
-        return config;
+        return Trinity::StringFormat("{}%", m_mapBuilder->currentPercentageDone());
     }
 
     /**************************************************************************/
