+++ src/tools/mmaps_generator/TerrainBuilder.cpp
@@ -29,10 +29,9 @@
 namespace MMAP
 {
     TerrainBuilder::TerrainBuilder(bool skipLiquid) : m_skipLiquid (skipLiquid){ }
-    TerrainBuilder::~TerrainBuilder() { }
 
     /**************************************************************************/
-    void TerrainBuilder::getLoopVars(Spot portion, int &loopStart, int &loopEnd, int &loopInc)
+    void TerrainBuilder::getLoopVars(Spot portion, int& loopStart, int& loopEnd, int& loopInc)
     {
         switch (portion)
         {
@@ -65,31 +64,31 @@ namespace MMAP
     }
 
     /**************************************************************************/
-    void TerrainBuilder::loadMap(uint32 mapID, uint32 tileX, uint32 tileY, MeshData &meshData)
+    void TerrainBuilder::loadMap(uint32 mapID, uint32 tileX, uint32 tileY, MeshData& meshData, VMAP::VMapManager2* vmapManager)
     {
-        if (loadMap(mapID, tileX, tileY, meshData, ENTIRE))
+        if (loadMap(mapID, tileX, tileY, meshData, vmapManager, ENTIRE))
         {
-            loadMap(mapID, tileX+1, tileY, meshData, LEFT);
-            loadMap(mapID, tileX-1, tileY, meshData, RIGHT);
-            loadMap(mapID, tileX, tileY+1, meshData, TOP);
-            loadMap(mapID, tileX, tileY-1, meshData, BOTTOM);
+            loadMap(mapID, tileX+1, tileY, meshData, vmapManager, LEFT);
+            loadMap(mapID, tileX-1, tileY, meshData, vmapManager, RIGHT);
+            loadMap(mapID, tileX, tileY+1, meshData, vmapManager, TOP);
+            loadMap(mapID, tileX, tileY-1, meshData, vmapManager, BOTTOM);
         }
     }
 
     /**************************************************************************/
-    bool TerrainBuilder::loadMap(uint32 mapID, uint32 tileX, uint32 tileY, MeshData &meshData, Spot portion)
+    bool TerrainBuilder::loadMap(uint32 mapID, uint32 tileX, uint32 tileY, MeshData& meshData, VMAP::VMapManager2* vmapManager, Spot portion)
     {
         std::string mapFileName = Trinity::StringFormat("maps/{:04}_{:02}_{:02}.map", mapID, tileY, tileX);
 
         FILE* mapFile = fopen(mapFileName.c_str(), "rb");
         if (!mapFile)
         {
-            int32 parentMapId = sMapStore[mapID].ParentMapID;
+            int32 parentMapId = vmapManager->getParentMapId(mapID);
             while (!mapFile && parentMapId != -1)
             {
                 mapFileName = Trinity::StringFormat("maps/{:04}_{:02}_{:02}.map", parentMapId, tileY, tileX);
                 mapFile = fopen(mapFileName.c_str(), "rb");
-                parentMapId = sMapStore[parentMapId].ParentMapID;
+                parentMapId = vmapManager->getParentMapId(mapID);
             }
         }
 
@@ -124,12 +123,9 @@ namespace MMAP
         }
 
         // data used later
-        uint8 holes[16][16][8];
-        memset(holes, 0, sizeof(holes));
-        uint16 liquid_entry[16][16];
-        memset(liquid_entry, 0, sizeof(liquid_entry));
-        map_liquidHeaderTypeFlags liquid_flags[16][16];
-        memset(liquid_flags, 0, sizeof(liquid_flags));
+        uint8 holes[16][16][8] = { };
+        uint16 liquid_entry[16][16] = { };
+        map_liquidHeaderTypeFlags liquid_flags[16][16] = { };
         G3D::Array<int> ltriangles;
         G3D::Array<int> ttriangles;
 
@@ -559,7 +555,7 @@ namespace MMAP
     }
 
     /**************************************************************************/
-    bool TerrainBuilder::isHole(int square, uint8 const holes[16][16][8])
+    bool TerrainBuilder::isHole(int square, uint8 const (&holes)[16][16][8])
     {
         int row = square / 128;
         int col = square % 128;
@@ -583,9 +579,8 @@ namespace MMAP
     }
 
     /**************************************************************************/
-    bool TerrainBuilder::loadVMap(uint32 mapID, uint32 tileX, uint32 tileY, MeshData &meshData)
+    bool TerrainBuilder::loadVMap(uint32 mapID, uint32 tileX, uint32 tileY, MeshData& meshData, VMAP::VMapManager2* vmapManager)
     {
-        std::unique_ptr<VMAP::VMapManager2> vmapManager = VMapFactory::CreateVMapManager();
         VMAP::LoadResult result = vmapManager->loadMap("vmaps", mapID, tileX, tileY);
         bool retval = false;
 
@@ -792,7 +787,7 @@ namespace MMAP
     }
 
     /**************************************************************************/
-    void TerrainBuilder::cleanVertices(G3D::Array<float> &verts, G3D::Array<int> &tris)
+    void TerrainBuilder::cleanVertices(G3D::Array<float>& verts, G3D::Array<int>& tris)
     {
         std::map<int, int> vertMap;
