+++ src/tools/mmaps_generator/TerrainBuilder.h
@@ -18,11 +18,15 @@
 #ifndef _MMAP_TERRAIN_BUILDER_H
 #define _MMAP_TERRAIN_BUILDER_H
 
-#include "PathCommon.h"
 #include "WorldModel.h"
 
 #include <G3D/Array.h>
 #include <G3D/Vector3.h>
+
+namespace VMAP
+{
+class VMapManager2;
+}
 
 enum class map_liquidHeaderTypeFlags : uint8;
 
@@ -91,11 +95,10 @@ namespace MMAP
     class TerrainBuilder
     {
         public:
-            TerrainBuilder(bool skipLiquid);
-            ~TerrainBuilder();
+            explicit TerrainBuilder(bool skipLiquid);
 
-            void loadMap(uint32 mapID, uint32 tileX, uint32 tileY, MeshData &meshData);
-            bool loadVMap(uint32 mapID, uint32 tileX, uint32 tileY, MeshData &meshData);
+            void loadMap(uint32 mapID, uint32 tileX, uint32 tileY, MeshData& meshData, VMAP::VMapManager2* vmapManager);
+            bool loadVMap(uint32 mapID, uint32 tileX, uint32 tileY, MeshData& meshData, VMAP::VMapManager2* vmapManager);
             void loadOffMeshConnections(uint32 mapID, uint32 tileX, uint32 tileY, MeshData& meshData, std::vector<OffMeshData> const& offMeshConnections);
 
             bool usesLiquids() const { return !m_skipLiquid; }
@@ -106,19 +109,16 @@ namespace MMAP
             static void copyVertices(std::vector<G3D::Vector3> const& source, G3D::Array<float>& dest);
             static void copyIndices(std::vector<VMAP::MeshTriangle> const& source, G3D::Array<int>& dest, int offset, bool flip);
             static void copyIndices(G3D::Array<int> const& source, G3D::Array<int>& dest, int offset);
-            static void cleanVertices(G3D::Array<float> &verts, G3D::Array<int> &tris);
+            static void cleanVertices(G3D::Array<float>& verts, G3D::Array<int>& tris);
         private:
             /// Loads a portion of a map's terrain
-            bool loadMap(uint32 mapID, uint32 tileX, uint32 tileY, MeshData &meshData, Spot portion);
+            bool loadMap(uint32 mapID, uint32 tileX, uint32 tileY, MeshData& meshData, VMAP::VMapManager2* vmapManager, Spot portion);
 
             /// Sets loop variables for selecting only certain parts of a map's terrain
-            void getLoopVars(Spot portion, int &loopStart, int &loopEnd, int &loopInc);
+            void getLoopVars(Spot portion, int& loopStart, int& loopEnd, int& loopInc);
 
             /// Controls whether liquids are loaded
             bool m_skipLiquid;
-
-            /// Load the map terrain from file
-            bool loadHeightMap(uint32 mapID, uint32 tileX, uint32 tileY, G3D::Array<float> &vertices, G3D::Array<int> &triangles, Spot portion);
 
             /// Get the vector coordinate for a specific position
             void getHeightCoord(int index, Grid grid, float xOffset, float yOffset, float* coord, float* v);
@@ -127,17 +127,13 @@ namespace MMAP
             void getHeightTriangle(int square, Spot triangle, int* indices, bool liquid = false);
 
             /// Determines if the specific position's triangles should be rendered
-            bool isHole(int square, uint8 const holes[16][16][8]);
+            bool isHole(int square, uint8 const (&holes)[16][16][8]);
 
             /// Get the liquid vector coordinate for a specific position
             void getLiquidCoord(int index, int index2, float xOffset, float yOffset, float* coord, float* v);
 
             /// Get the liquid type for a specific position
             map_liquidHeaderTypeFlags getLiquidType(int square, map_liquidHeaderTypeFlags const (&liquid_type)[16][16]);
-
-            // hide parameterless and copy constructor
-            TerrainBuilder() = delete;
-            TerrainBuilder(TerrainBuilder const& tb) = delete;
     };
 }
